+++ app/src/main/java/edu/javeriana/fixup/data/repository/ProfileRepository.kt
@@ -52,7 +52,7 @@ class ProfileRepository @Inject constructor(
             val reviewToSave = edu.javeriana.fixup.ui.model.ReviewModel(
                 userId = "1", // ID fijo quemado
                 userName = user?.displayName ?: user?.email?.substringBefore("@") ?: "Usuario Prueba",
-                rating = rating,
+                rating = rating.toDouble(),
                 comment = comment,
                 date = java.text.SimpleDateFormat("dd/MM/yyyy", java.util.Locale.getDefault()).format(java.util.Date())
             )
@@ -68,7 +68,7 @@ class ProfileRepository @Inject constructor(
             val reviewToUpdate = edu.javeriana.fixup.ui.model.ReviewModel(
                 id = reviewId,
                 userId = "1", // ID fijo quemado
-                rating = rating,
+                rating = rating.toDouble(),
                 comment = comment
             )
             val updatedReview = profileDataSource.updateReview(reviewId, reviewToUpdate)
